playwright.config.ts: Remove unused environment configuration and baseURL settings
tests/e2e: Add regression tests for each layer and core functionaly
