+++ playwright.config.ts
@@ -9,9 +9,6 @@ import { defineConfig, devices } from '@playwright/test';
  * Read environment variables from file.
  * https://github.com/motdotla/dotenv
  */
-// import dotenv from 'dotenv';
-// import path from 'path';
-// dotenv.config({ path: path.resolve(__dirname, '.env') });
 
 /**
  * See https://playwright.dev/docs/test-configuration.
@@ -29,15 +26,6 @@ export default defineConfig({
   /* Reporter to use. See https://playwright.dev/docs/test-reporters */
   reporter: 'html',
   /* Shared settings for all the projects below. See https://playwright.dev/docs/api/class-testoptions. */
-  use: {
-    /* Base URL to use in actions like `await page.goto('/')`. */
-    baseURL: 'https://localhost:8080',
-    ignoreHTTPSErrors: true,
-
-    /* Collect trace when retrying the failed test. See https://playwright.dev/docs/trace-viewer */
-    trace: 'on-first-retry',
-  },
-
   /* Configure projects for major browsers */
   projects: [
     {
@@ -45,10 +33,10 @@ export default defineConfig({
       use: { ...devices['Desktop Chrome'] },
     },
 
-    // {
-    //   name: 'firefox',
-    //   use: { ...devices['Desktop Firefox'] },
-    // },
+    {
+      name: 'firefox',
+      use: { ...devices['Desktop Firefox'] },
+    },
 
     // {
     //   name: 'webkit',
@@ -76,23 +64,4 @@ export default defineConfig({
     // },
   ],
 
-  /* Run your local dev server before starting the tests */
-  webServer: [
-    {
-      command: 'node scripts/setup-certs.js && npx http-server -p 8080 -S -C cert.pem -K key.pem',
-      url: 'https://localhost:8080',
-      reuseExistingServer: !process.env.CI,
-      ignoreHTTPSErrors: true,
-      stdout: 'ignore',
-      stderr: 'pipe',
-    },
-    {
-      command: 'node scripts/setup-certs.js && npx http-server demo-content -p 8081 -S -C cert.pem -K key.pem',
-      url: 'https://localhost:8081',
-      reuseExistingServer: !process.env.CI,
-      ignoreHTTPSErrors: true,
-      stdout: 'ignore',
-      stderr: 'pipe',
-    },
-  ],
 });
